requirements.txt
asgiref==3.2.10
certifi==2022.5.18
charset-normalizer==2.0.12
Django==3.1
django-filter==21.1
djangorestframework==3.13.1
idna==3.3
Markdown==3.3.7
protobuf==3.20.1
psycopg2==2.9.3
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
uwsgi
drf-yasg
django-rest-swagger
django-cors-headers
